Remove no-longer-necessary FFI bindings, fix hash-object newline escaping
