+++ .vscode/settings.json
@@ -8,6 +8,7 @@
   "eslint.packageManager": "pnpm",
   "debug.javascript.unmapMissingSources": true,
   "go.lintTool": "golangci-lint",
+  "go.buildTags": "rust",
   "files.associations": {
     "libturbo.h": "c",
     "turbo.json": "jsonc"
